+++ src/Road.h
@@ -16,13 +16,13 @@ class Road {
 	bool twoWay; /**<private member that specifies if the road is twoWay, true if so or false if not>  */
 public:
 	Road();
-	Road(unsigned int id, bool twoWay);
-	Road(unsigned int id, const string &name, bool twoWay);
+	Road(unsigned long id, bool twoWay);
+	Road(unsigned long id, const string &name, bool twoWay);
 	void setName(const string &name);
 	const string& getName() const;
 	void setTwoWay(bool twoWay);
 	bool getTwoWay() const;
-	unsigned int getId() const;
+	unsigned long getId() const;
 	bool operator == (const Road &p2) const;
 	friend ostream & operator << (ostream &os, Road &p);
 };
